Move simd_input and associated functions to their own header
